chore: move helper functions to designated file
